requirements.txt

flask
flask-cors
langchain
langchain-community
langchain-google-genai
langchain-pinecone
pinecone-client
sentence-transformers
pypdf
python-dotenv
huggingface-hub
